Push data fetching up
Wanted better cohesion between state declarations and the data fetching.
